+++ migrations/000016_places_search_cache.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE places_search_cache;